+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fixed "context canceled" warnings for typing action by using detached context with timeout
+- Fixed bot token leaking in error log messages - now sanitized as `[REDACTED]`
+
 ## [0.1.3] - 2026-01-01
 
 ### Changed